refactor: use modularized lodash.* packages
