+++ @types/lodash.get.d.ts
@@ -1,0 +1,4 @@
+declare module "lodash.get" {
+  import { get } from "lodash";
+  export default get;
+}
