+++ @types/lodash.uniq.d.ts
@@ -1,0 +1,4 @@
+declare module "lodash.uniq" {
+  import { uniq } from "lodash";
+  export default uniq;
+}
