+++ package.json
@@ -43,7 +43,8 @@
   "dependencies": {
     "@babel/runtime": "^7.0.0",
     "jsonld": "^1.5.0",
-    "lodash": "^4.17.15"
+    "lodash.get": "^4.4.2",
+    "lodash.uniq": "^4.5.0"
   },
   "scripts": {
     "test": "jest",
