+++ src/hydra/fetchResource.ts
@@ -1,4 +1,4 @@
-import { get } from "lodash";
+import get from "lodash.get";
 import fetchJsonLd from "./fetchJsonLd";
 
 export default (
