+++ src/swagger/handleJson.ts
@@ -1,4 +1,5 @@
-import { get, uniq } from "lodash";
+import get from "lodash.get";
+import uniq from "lodash.uniq";
 import { OpenAPIV2 } from "openapi-types";
 import { Field } from "../Field";
 import { Resource } from "../Resource";
