+++ tsconfig.json
@@ -8,6 +8,7 @@
     "importHelpers": true,
     "strict": true,
     "moduleResolution": "node",
-    "esModuleInterop": true
+    "esModuleInterop": true,
+    "typeRoots": ["node_modules/@types", "@types"]
   }
 }
